# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Management API",
        "description": "API documentation for the Event Management System. This system allows managing yearly events with all their associated content including summaries, themes, programmes, speakers, sponsors, FAQs, media, galleries, and attendances.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/topics": {
            "get": {
                "tags": ["Topics"],
                "summary": "Get all topics",
                "description": "Returns a list of all topics with their speakers",
                "operationId": "getTopics",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "query",
                        "description": "Filter topics by event UUID",
                        "required": false,
                        "schema": {"type": "string", "format": "uuid"}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {"type": "boolean", "example": true},
                                        "data": {
                                            "type": "array",
                                            "items": {"$ref": "#/components/schemas/Topic"}
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Topics"],
                "summary": "Create a new topic",
                "description": "Creates a new topic for an event",
                "operationId": "createTopic",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {"$ref": "#/components/schemas/TopicRequest"}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Topic created successfully",
                        "content": {
                            "application/json": {
                                "schema": {"$ref": "#/components/schemas/Topic"}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/topics/{id}": {
            "get": {
                "tags": ["Topics"],
                "summary": "Get topic by ID",
                "description": "Returns a single topic with all speakers",
                "operationId": "getTopicById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Topic UUID",
                        "required": true,
                        "schema": {"type": "string", "format": "uuid"}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {"$ref": "#/components/schemas/Topic"}
                            }
                        }
                    },
                    "404": {
                        "description": "Topic not found"
                    }
                }
            },
            "put": {
                "tags": ["Topics"],
                "summary": "Update a topic",
                "description": "Updates an existing topic",
                "operationId": "updateTopic",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Topic UUID",
                        "required": true,
                        "schema": {"type": "string", "format": "uuid"}
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {"$ref": "#/components/schemas/TopicRequest"}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Topic updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {"$ref": "#/components/schemas/Topic"}
                            }
                        }
                    },
                    "404": {
                        "description": "Topic not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": ["Topics"],
                "summary": "Delete a topic",
                "description": "Deletes a topic (speakers will have topic_id set to null)",
                "operationId": "deleteTopic",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Topic UUID",
                        "required": true,
                        "schema": {"type": "string", "format": "uuid"}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Topic deleted successfully"
                    },
                    "404": {
                        "description": "Topic not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Topic": {
                "type": "object",
                "properties": {
                    "id": {"type": "string", "format": "uuid", "example": "8a2c7b3d-6e4f-5c1a-2b9e-3f4a5b6c7d8e"},
                    "event_id": {"type": "string", "format": "uuid", "example": "9b3d8f4e-7c1a-4d2b-9e5f-1a2b3c4d5e6f"},
                    "title": {"type": "string", "example": "The Future of Artificial Intelligence"},
                    "topic_date": {"type": "string", "format": "date", "example": "2026-11-10"},
                    "content": {"type": "string", "example": "Exploring how AI is transforming industries and creating new opportunities..."},
                    "topic_picture": {"type": "string", "example": "/images/topics/ai-future.jpg"},
                    "order": {"type": "integer", "example": 1},
                    "created_at": {"type": "string", "format": "date-time"},
                    "updated_at": {"type": "string", "format": "date-time"},
                    "speakers": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {"type": "string", "format": "uuid"},
                                "name": {"type": "string"},
                                "title": {"type": "string"},
                                "organization": {"type": "string"}
                            }
                        }
                    }
                }
            },
            "TopicRequest": {
                "type": "object",
                "required": ["event_id", "title", "topic_date"],
                "properties": {
                    "event_id": {"type": "string", "format": "uuid", "example": "9b3d8f4e-7c1a-4d2b-9e5f-1a2b3c4d5e6f"},
                    "title": {"type": "string", "example": "The Future of Artificial Intelligence"},
                    "topic_date": {"type": "string", "format": "date", "example": "2026-11-10"},
                    "content": {"type": "string", "example": "Exploring how AI is transforming industries..."},
                    "topic_picture": {"type": "string", "example": "/images/topics/ai-future.jpg"},
                    "order": {"type": "integer", "example": 1}
                }
            }
        }
    },
...
}